requirements.txt
Django==2.2.7
django-modeltranslation==0.14.4
Pillow==7.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
